ml/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
python-dotenv
groq
